Join Operations in MapReduce
Customer Life Time Value (LTV) is estimated for each client of a small medical enterprise in Christchurch New Zealand that offers specialised general consultation services. The LTV is determined from customer and transaction details dataset. Hadoop MapReduce platform was used to map and reduce the dataset to required key-value object. 

The key disadvantage of using MapReduce for join operations is the high compute power for executing the mapper operation. For join operations on a larger dataset, Hive as one of the hadoop ecosystem component is preferred. It shares similar logic with SQL query.